feat : 유저 사진 업로드 기능 개발

See merge request s13-fintech-finance-sub1/S13P21A408!18

+++ Backend/src/main/java/com/mango/backend/domain/user/dto/response/MyInfoResponse.java
@@ -9,7 +9,7 @@ public record MyInfoResponse(
     String email,
     String nickname,
     LocalDate birthDate,
-    Byte age,
+    Integer age,
     String gender,
     String latitude,
     String longitude,
